Extracts mouse interaction as separate fixture class
- ToolTestFixture does not inherit from testing::Test so that it can
  be used together with DisplayTestFixture. Multiple inheritance of
  test fixtures does not work properly.
