+++ rviz_default_plugins/test/rviz_default_plugins/displays/display_test_fixture.hpp
@@ -58,7 +58,7 @@ public:
 
   DisplayTestFixture();
 
-  ~DisplayTestFixture();
+  ~DisplayTestFixture() override;
 
   static void TearDownTestCase();
